requirements_gpu.txt
dm-tree==0.1.1
cleverhans==4.0.0
tensorflow-gpu==2.4.0
cntk==2.5
pillow==8.0
urllib3
opencv-python
